Align sequences using SPOA (and optionally get the consensus)

Align sequences using SPOA (and optionally get the consensus)

spoaAlign(
  seq,
  m = 5L,
  n = -4L,
  g = -8L,
  e = g,
  q = g,
  c = e,
  algorithm = c("local", "global", "semi.global"),
  w = 1,
  ...
)

spoaConsensus(
  seq,
  m = 5L,
  n = -4L,
  g = -8L,
  e = g,
  q = g,
  c = e,
  algorithm = c("local", "global", "semi.global"),
  w = 1L,
  ...
)

Arguments

seq
(character vector or Biostrings::XStringSet) sequences to align

m
(non-negative integer) score for a match. Default: 5L

n
(non-positive integer) score for a mismatch. Default: -4L

g
(non-positive integer) gap opening penalty. Default: -8L

e
(non-positive integer) gap extension penalty. Default: same value as g

q
(non-positive integer) second gap opening penalty. Default: same value as g

c
(non-positive integer) second gap extension penalty. Default: same value as e

algorithm
(character string) alignment mode; one of "local" (Smith-Watterman), "global" (Needleman-Wunsch), or "semi.global" (Overlap). Default: "local"

w
(integer vector) weights for the sequences. Default: 1L

...
additional parameters passed to methods

Value

For spoaConsensus(), either a character string or the appropriate Biostrings::XString, depending on the class of seq.

For spoaAlign(), either a character vector or a Biostrings::MultipleAlignment, depending on the class of seq. If seq is a BStringSet (i.e., an XStringSet which is not specifically DNA, RNA, or AA) then the result is also a BStringSet, since there is no corresponding "BMultipleAlignment" class.

Details

Gap penalties

The general (convex) gap penalty formula is:

min(g + (i - i) * e, q + (i - 1) * c)

For q == g and c == e (as is the case unless q or c is explicitly set), this simplifies to the affine gap penalty:

g + (i - 1) * e

If, additionally, e == g (the default), then this is the linear gap penalty:

g * i

Weighting

Assigning a weight w to a sequence is equivalent to including that sequence w times; this is primarily useful for dereplicated sequences, where all sequences are unique, and the weight represents their multiplicity.

Examples

sequences <- c(
    "CATAAAAGAACGTAGGTCGCCCGTCCGTAACCTGTCGGATCACCGGAAAGGACCCGTAAAGTGATAATGAT",
    "ATAAAGGCAGTCGCTCTGTAAGCTGTCGATTCACCGGAAAGATGGCGTTACCACGTAAAGTGATAATGATTAT",
    "ATCAAAGAACGTGTAGCCTGTCCGTAATCTAGCGCATTTCACACGAGACCCGCGTAATGGG",
    "CGTAAATAGGTAATGATTATCATTACATATCACAACTAGGGCCGTATTAATCATGATATCATCA",
    "GTCGCTAGAGGCATCGTGAGTCGCTTCCGTACCGCAAGGATGACGAGTCACTTAAAGTGATAAT",
    "CCGTAACCTTCATCGGATCACCGGAAAGGACCCGTAAATAGACCTGATTATCATCTACAT"
)
spoaAlign(sequences)

#> [1] "CATAAAAGAACG------T---------AGGTCGCCCGTCCGTAACC-T-GTCG-G-ATCACCGG-AA-A--G-G--A-CC-CGTAAAGTGATAATGAT-------------"
#> [2] "------------------------ATAAAGGCAGTCGCTCTGTAAGC-T-GTCG-A-TTCACCGGAAAGATGGCGTTA-CCACGTAAAGTGATAATGATTAT----------"
#> [3] "-ATCAAAGAACG------T-----------GTAGCCTGTCCGTAATC-T-AGCGCATTTCACACG--AGA---C-----CCGCGTAATGGG---------------------"
#> [4] "------------CGTAAAT---------AGGTAAT-GAT-TATCAT--T-A-C--ATATCACAAC--T-A--G-G----GC-CGT-AT-TAATCATGA-TATCATCA-----"
#> [5] "-------------------GTCGC-TAGAGGCA-T-CGTGAGTC-GC-T--TCC-G--T-ACCGCAAGGATGACG--AGTCACTTAAAGTGATAAT----------------"
#> [6] "---------------------------------------CCGTAACCTTCATCG-G-ATCACCGG-AA-A--G-G--A-CC-CGTAAA-TAGACCTGATTATCATC-TACAT"

spoaConsensus(sequences)

#> [1] "CATCAAAGAACGTAGGTAGCCCGTCCGTAACCTGTCGGATCACCGGAAAGAGGACCCGTAAAGTGATAATGATTATCATCTACAT"